Code clean up.

+++ core/src/main/java/com/orientechnologies/orient/core/metadata/security/ORule.java
@@ -305,7 +305,7 @@ public class ORule implements Serializable {
       return null;
     }
 
-    final byte mask = (byte) operation;
+    final byte mask = operation;
 
     return (ac.byteValue() & mask) == mask;
   }
