+++ core/src/main/java/com/orientechnologies/orient/core/storage/cluster/v0/OClusterPositionMapV0.java
@@ -98,7 +98,7 @@ public class OClusterPositionMapV0 extends OClusterPositionMap {
       }
 
       final long index = bucket.add(pageIndex, recordPosition);
-      return index + cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES;
+      return index + (long) cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES;
     } finally {
       cacheEntry.close();
     }
@@ -132,7 +132,7 @@ public class OClusterPositionMapV0 extends OClusterPositionMap {
       }
 
       final long index = bucket.allocate();
-      return index + cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES;
+      return index + (long) cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES;
     } finally {
       cacheEntry.close();
     }
@@ -232,7 +232,7 @@ public class OClusterPositionMapV0 extends OClusterPositionMap {
         } else {
           int entriesCount = 0;
           final long startIndex =
-              cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES + index;
+              (long) cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES + index;
 
           result = new OClusterPositionEntry[resultSize];
           for (int i = 0; i < resultSize; i++) {
@@ -293,7 +293,7 @@ public class OClusterPositionMapV0 extends OClusterPositionMap {
         } else {
           int entriesCount = 0;
           final long startIndex =
-              cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES + index;
+              (long) cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES + index;
 
           result = new long[resultSize];
           for (int i = 0; i < resultSize; i++) {
@@ -363,7 +363,7 @@ public class OClusterPositionMapV0 extends OClusterPositionMap {
         int entriesCount = 0;
 
         final long startPosition =
-            cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES;
+            (long) cacheEntry.getPageIndex() * OClusterPositionMapBucket.MAX_ENTRIES;
         result = new long[resultSize];
 
         for (int i = 0; i < resultSize; i++) {
